services/ml/requirements.txt
numpy==1.26.4
pillow==10.4.0
opencv-python-headless==4.10.0.84
faiss-cpu==1.8.0.post1
torch==2.4.1
open_clip_torch==2.26.1
pandas==2.2.2
scikit-learn==1.5.1
requests==2.32.3
python-dotenv==1.0.1
